backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.0.1
httpx==0.28.1
pytest==9.0.2
PyYAML==6.0.2
tencentcloud-sdk-python>=3.0.0
